requirements.txt
torch>=1.12.0
numpy>=1.20.0
gym>=0.25.0
matplotlib>=3.5.0
PyYAML>=6.0
pygame>=2.1.0
box2d-py>=2.3.0